% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_summary_of_indicators.R
\name{load_summary_of_indicators}
\alias{load_summary_of_indicators}
\title{Load indicator summary
This function loads the summary of indicator table data from the \code{inst/extdata} directory.}
\usage{
load_summary_of_indicators()
}
\value{
A data frame containing the example data.
}
\description{
Load indicator summary
This function loads the summary of indicator table data from the \code{inst/extdata} directory.
}
\examples{
summary.df <- load_summary_of_indicators()
head(summary.df)

}
